refactor(model): model 을 사용하는 클래스 내부 import 리팩토링
