+++ src/main/java/racingcar/InputDTO.java
@@ -1,6 +1,7 @@
 package racingcar;
 
 import java.util.List;
+import racingcar.model.RacingCar;
 
 public record InputDTO(List<String> carNames, Long tryNums) {
     public List<RacingCar> to() {
