+++ src/main/java/racingcar/controller/RandomRaceController.java
@@ -2,7 +2,7 @@ package racingcar.controller;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
-import racingcar.RacingCar;
+import racingcar.model.RacingCar;
 import racingcar.view.RaceProgressView;
 
 public class RandomRaceController {
