+++ src/test/java/racingcar/RacingCarTest.java
@@ -3,6 +3,7 @@ package racingcar;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
+import racingcar.model.RacingCar;
 
 class RacingCarTest {
