+++ src/test/java/racingcar/RandomRaceControllerTest.java
@@ -7,6 +7,7 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import racingcar.controller.RandomRaceController;
+import racingcar.model.RacingCar;
 
 class RandomRaceControllerTest {
     private RandomRaceController randomRaceController;
